restructure the tests driven by data now
